eliminate warnings, compiles for fedora 29
