+++ src/common/CMakeLists.txt
@@ -3,6 +3,10 @@ aux_source_directory(. COMMON_SRC)
 
 file(GLOB COMMON_INC "*.h")
 
+if (${CMAKE_CXX_COMPILER_ID} STREQUAL "GNU")
+  # must compile with -fPIC
+  add_compile_options(-fPIC)
+endif()
 add_library(common ${VM_SHARED_OR_STATIC})
 target_sources(common PRIVATE ${COMMON_SRC})
 target_include_directories(common PRIVATE "${CMAKE_SOURCE_DIR}/lib/3rdparty")
